Working WIP with wrapper class embedded in codex_cytokit
